refactor(core): move ledger translate functions
