+++ core/api/src/debug/clean-pending-payments-for-dead-node.ts
@@ -13,7 +13,8 @@ import { isUp } from "@/services/lnd/health"
 import { lndsConnect } from "@/services/lnd/auth"
 
 import { setupMongoConnection } from "@/services/mongodb"
-import { LedgerService, translateToLedgerTx } from "@/services/ledger"
+import { LedgerService } from "@/services/ledger"
+import { translateToLedgerTx } from "@/services/ledger/translate"
 import * as LedgerFacade from "@/services/ledger/facade"
 import { LockService } from "@/services/lock"
 import { baseLogger } from "@/services/logger"
